+++ setuptools/tests/files.py
@@ -6,10 +6,13 @@ import pkg_resources.py31compat
 
 def build_files(file_defs, prefix=""):
     """
-    Build a set of files/directories, as described by the file_defs dictionary.
+    Build a set of files/directories, as described by the
+    file_defs dictionary.
 
-    Each key/value pair in the dictionary is interpreted as a filename/contents
-    pair. If the contents value is a dictionary, a directory is created, and the
+    Each key/value pair in the dictionary is interpreted as
+    a filename/contents
+    pair. If the contents value is a dictionary, a directory
+    is created, and the
     dictionary interpreted as the files within it, recursively.
 
     For example: