+++ setuptools/tests/test_config.py
@@ -1,10 +1,18 @@
+# -*- coding: UTF-8 -*-
+from __future__ import unicode_literals
+
 import contextlib
 import pytest
+
 from distutils.errors import DistutilsOptionError, DistutilsFileError
 from mock import patch
 from setuptools.dist import Distribution, _Distribution
 from setuptools.config import ConfigHandler, read_configuration
+from setuptools.extern.six.moves.configparser import InterpolationMissingOptionError
+from setuptools.tests import is_ascii
 from . import py2_only, py3_only
+from .textwrap import DALS
+
 
 class ErrConfigHandler(ConfigHandler):
     """Erroneous handler. Fails to implement required methods."""
@@ -16,12 +24,12 @@ def make_package_dir(name, base_dir, ns=False):
         dir_package = dir_package.mkdir(dir_name)
     init_file = None
     if not ns:
-      init_file = dir_package.join('__init__.py')
-      init_file.write('')
+        init_file = dir_package.join('__init__.py')
+        init_file.write('')
     return dir_package, init_file
 
 
-def fake_env(tmpdir, setup_cfg, setup_py=None, package_path='fake_package'):
+def fake_env(tmpdir, setup_cfg, setup_py=None, encoding='ascii', package_path='fake_package'):
 
     if setup_py is None:
         setup_py = (
@@ -31,7 +39,7 @@ def fake_env(tmpdir, setup_cfg, setup_py=None, package_path='fake_package'):
 
     tmpdir.join('setup.py').write(setup_py)
     config = tmpdir.join('setup.cfg')
-    config.write(setup_cfg)
+    config.write(setup_cfg.encode(encoding), mode='wb')
 
     package_dir, init_file = make_package_dir(package_path, tmpdir)
 
@@ -145,6 +153,24 @@ class TestMetadata:
             assert metadata.keywords == ['one', 'two']
             assert metadata.download_url == 'http://test.test.com/test/'
             assert metadata.maintainer_email == 'test@test.com'
+
+    def test_license_cfg(self, tmpdir):
+        fake_env(
+            tmpdir,
+            DALS("""
+            [metadata]
+            name=foo
+            version=0.0.1
+            license=Apache 2.0
+            """)
+        )
+
+        with get_dist(tmpdir) as dist:
+            metadata = dist.metadata
+
+            assert metadata.name == "foo"
+            assert metadata.version == "0.0.1"
+            assert metadata.license == "Apache 2.0"
 
     def test_file_mixed(self, tmpdir):
 
@@ -288,7 +314,7 @@ class TestMetadata:
         tmpdir.join('fake_package', 'version.txt').write('1.2.3\n4.5.6\n')
         with pytest.raises(DistutilsOptionError):
             with get_dist(tmpdir) as dist:
-                _ = dist.metadata.version
+                dist.metadata.version
 
     def test_version_with_package_dir_simple(self, tmpdir):
 
@@ -391,6 +417,89 @@ class TestMetadata:
         with get_dist(tmpdir) as dist:
             assert set(dist.metadata.classifiers) == expected
 
+    def test_deprecated_config_handlers(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '[metadata]\n'
+            'version = 10.1.1\n'
+            'description = Some description\n'
+            'requires = some, requirement\n'
+        )
+
+        with pytest.deprecated_call():
+            with get_dist(tmpdir) as dist:
+                metadata = dist.metadata
+
+                assert metadata.version == '10.1.1'
+                assert metadata.description == 'Some description'
+                assert metadata.requires == ['some', 'requirement']
+
+    def test_interpolation(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '[metadata]\n'
+            'description = %(message)s\n'
+        )
+        with pytest.raises(InterpolationMissingOptionError):
+            with get_dist(tmpdir):
+                pass
+
+    skip_if_not_ascii = pytest.mark.skipif(not is_ascii, reason='Test not supported with this locale')
+
+    @skip_if_not_ascii
+    def test_non_ascii_1(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '[metadata]\n'
+            'description = éàïôñ\n',
+            encoding='utf-8'
+        )
+        with pytest.raises(UnicodeDecodeError):
+            with get_dist(tmpdir):
+                pass
+
+    def test_non_ascii_2(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '# -*- coding: invalid\n'
+        )
+        with pytest.raises(LookupError):
+            with get_dist(tmpdir):
+                pass
+
+    def test_non_ascii_3(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '\n'
+            '# -*- coding: invalid\n'
+        )
+        with get_dist(tmpdir):
+            pass
+
+    @skip_if_not_ascii
+    def test_non_ascii_4(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '# -*- coding: utf-8\n'
+            '[metadata]\n'
+            'description = éàïôñ\n',
+            encoding='utf-8'
+        )
+        with get_dist(tmpdir) as dist:
+            assert dist.metadata.description == 'éàïôñ'
+
+    @skip_if_not_ascii
+    def test_non_ascii_5(self, tmpdir):
+        fake_env(
+            tmpdir,
+            '# vim: set fileencoding=iso-8859-15 :\n'
+            '[metadata]\n'
+            'description = éàïôñ\n',
+            encoding='iso-8859-15'
+        )
+        with get_dist(tmpdir) as dist:
+            assert dist.metadata.description == 'éàïôñ'
+
 
 class TestOptions:
 
@@ -414,7 +523,7 @@ class TestOptions:
             'tests_require = mock==0.7.2; pytest\n'
             'setup_requires = docutils>=0.3; spack ==1.1, ==1.3; there\n'
             'dependency_links = http://some.com/here/1, '
-                'http://some.com/there/2\n'
+            'http://some.com/there/2\n'
             'python_requires = >=1.0, !=2.8\n'
             'py_modules = module1, module2\n'
         )
@@ -622,7 +731,7 @@ class TestOptions:
         dir_sub_two, _ = make_package_dir('sub_two', dir_package, ns=True)
 
         with get_dist(tmpdir) as dist:
-            assert set(dist.packages) == { 
+            assert set(dist.packages) == {
                 'fake_package', 'fake_package.sub_two', 'fake_package.sub_one'
             }
 
@@ -674,7 +783,7 @@ class TestOptions:
             tmpdir,
             '[options.entry_points]\n'
             'group1 = point1 = pack.module:func, '
-                '.point2 = pack.module2:func_rest [rest]\n'
+            '.point2 = pack.module2:func_rest [rest]\n'
             'group2 = point3 = pack.module:func2\n'
         )
 
@@ -720,7 +829,10 @@ class TestOptions:
             ]
             assert sorted(dist.data_files) == sorted(expected)
 
+
 saved_dist_init = _Distribution.__init__
+
+
 class TestExternalSetters:
     # During creation of the setuptools Distribution() object, we call
     # the init of the parent distutils Distribution object via