+++ setuptools/tests/test_namespaces.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import, unicode_literals
 
-import os
 import sys
 import subprocess
 
@@ -12,7 +11,7 @@ from setuptools.command import test
 
 class TestNamespaces:
 
-    @pytest.mark.xfail(
+    @pytest.mark.skipif(
         sys.version_info < (3, 5),
         reason="Requires importlib.util.module_from_spec",
     )