+++ setuptools/tests/test_virtualenv.py
@@ -57,9 +57,6 @@ def test_pip_upgrade_from_source(virtualenv):
     Check pip can upgrade setuptools from source.
     """
     dist_dir = virtualenv.workspace
-    if sys.version_info < (2, 7):
-        # Python 2.6 support was dropped in wheel 0.30.0.
-        virtualenv.run('pip install -U "wheel<0.30.0"')
     # Generate source distribution / wheel.
     virtualenv.run(' && '.join((
         'cd {source}',
@@ -137,3 +134,14 @@ def test_test_command_install_requirements(bare_virtualenv, tmpdir):
         'python setup.py test -s test',
     )).format(tmpdir=tmpdir))
     assert tmpdir.join('success').check()
+
+
+def test_no_missing_dependencies(bare_virtualenv):
+    """
+    Quick and dirty test to ensure all external dependencies are vendored.
+    """
+    for command in ('upload',):  # sorted(distutils.command.__all__):
+        bare_virtualenv.run(' && '.join((
+            'cd {source}',
+            'python setup.py {command} -h',
+        )).format(command=command, source=SOURCE_DIR))